feat(dashboard): introduce TriggerStepperModal for streamlined trigger creation and editing process, enhancing user experience and modularity
