+++ frontend/src/components/dashboard/Triggers.vue
@@ -21,76 +21,26 @@
     </template>
   </Card>
 
-  <StepperModal
-    ref="triggerStepper"
-    size="lg"
-    :steps="stepperSteps"
-    :submit-text="stepperSubmitText"
-    @submit="save"
-  >
-    <template #title>
-      <h5 class="modal-title">{{ editingId ? "Edit trigger" : "Create trigger" }}</h5>
-    </template>
-    <template #step-1>
-      <BasicForm v-model="triggerForm" :fields="settingsFields" />
-    </template>
-    <template #step-2>
-      <BasicForm v-model="triggerForm" :fields="eventFields" />
-      <hr v-if="eventConfigFields.length" />
-      <BasicForm
-        v-if="eventConfigFields.length"
-        v-model="eventData"
-        :fields="eventConfigFields"
-      />
-    </template>
-    <template #step-3>
-      <BasicForm v-model="triggerForm" :fields="actionSelectFields" />
-      <hr v-if="actionConfigFields.length" />
-      <BasicForm
-        v-if="actionConfigFields.length"
-        v-model="actionData"
-        :fields="actionConfigFields"
-      />
-    </template>
-  </StepperModal>
-
+  <TriggerStepperModal ref="triggerStepper" />
   <ConfirmModal ref="deleteModal" />
-
-  <BasicModal
-    v-if="viewFormData"
-    ref="viewModal"
-    name="trigger-view"
-    size="lg"
-  >
-    <template #title>
-      {{ viewModalTitle }}
-    </template>
-    <template #body>
-      <BasicForm :model-value="viewFormData" :fields="viewFormSchema" />
-    </template>
-  </BasicModal>
 </template>
 
 <script>
 import Card from "@/basic/dashboard/card/Card.vue";
 import BasicTable from "@/basic/Table.vue";
 import BasicButton from "@/basic/Button.vue";
-import StepperModal from "@/basic/modal/StepperModal.vue";
-import BasicForm from "@/basic/Form.vue";
-import BasicModal from "@/basic/Modal.vue";
 import ConfirmModal from "@/basic/modal/ConfirmModal.vue";
+import TriggerStepperModal from "./triggers/TriggerStepperModal.vue";
 
 export default {
   name: "DashboardTriggers",
-  subscribeTable: ["trigger", "trigger_event", "trigger_action", "project", "template", "submission"],
+  subscribeTable: ["trigger", "trigger_event", "trigger_action"],
   components: {
     Card,
     BasicTable,
     BasicButton,
-    StepperModal,
-    BasicForm,
-    BasicModal,
     ConfirmModal,
+    TriggerStepperModal,
   },
   data() {
     return {
@@ -112,7 +62,6 @@ export default {
           action: "view",
           handler: "viewModal",
           options: { iconOnly: true, specifiers: { "btn-outline-secondary": true } },
-          modal: "view",
         },
         {
           icon: "pencil-square",
@@ -139,121 +88,6 @@ export default {
         action: "toggleEnabled",
         socketEvent: "triggerUpdate",
       },
-      stepperSteps: [
-        { title: "Trigger info" },
-        { title: "Event" },
-        { title: "Action" },
-      ],
-      stepperSubmitText: "Save",
-      settingsFormSchema: [
-        {
-          key: "name",
-          label: "Name",
-          type: "text",
-          required: true,
-          help: "Display name for this trigger rule in the Triggers dashboard.",
-        },
-        {
-          key: "description",
-          label: "Description",
-          type: "textarea",
-          required: true,
-          help: "Short note for admins describing when and why this trigger runs.",
-        },
-        {
-          key: "projectId",
-          label: "Scope to project",
-          type: "select",
-          required: true,
-          help: "Limits this trigger to the selected project. Event filters and assignment options use this project.",
-          optionsSource: {
-            table: "project",
-            labelKey: "name",
-            valueKey: "id",
-          },
-        },
-        {
-          key: "maxRetries",
-          label: "Max retries",
-          type: "number",
-          min: 0,
-          required: true,
-          help: "How many times a failed run may be retried from the trigger logs before it stays failed.",
-        },
-        {
-          key: "parallelLimit",
-          label: "Parallel limit",
-          type: "number",
-          min: 1,
-          required: true,
-          help: "Maximum number of executions of this trigger that may run at the same time.",
-        },
-        {
-          key: "timeout",
-          label: "Timeout (seconds)",
-          type: "number",
-          min: 1,
-          required: true,
-          help: "Maximum seconds a single execution may run before it is treated as timed out.",
-        },
-      ],
-      eventField: {
-        key: "triggerEventId",
-        label: "When (event)",
-        type: "select",
-        required: true,
-        optionsSource: {
-          table: "trigger_event",
-          labelKey: "configuration.label",
-          nameKey: "name",
-          valueKey: "id",
-          filter: { enabled: true },
-        },
-      },
-      actionField: {
-        key: "triggerActionId",
-        label: "Then (action)",
-        type: "select",
-        required: true,
-        optionsSource: {
-          table: "trigger_action",
-          labelKey: "configuration.label",
-          nameKey: "name",
-          valueKey: "id",
-          filter: { enabled: true },
-          compatibleWithEvent: true,
-        },
-      },
-      viewModalTitleTemplate: "Trigger: {name}",
-      viewFormSchema: [
-        { key: "name", label: "Name", type: "text", readOnly: true },
-        { key: "eventLabel", label: "Event", type: "text", readOnly: true },
-        { key: "actionLabel", label: "Action", type: "text", readOnly: true },
-        { key: "maxRetries", label: "Max retries", type: "text", readOnly: true },
-        { key: "parallelLimit", label: "Parallel limit", type: "text", readOnly: true },
-        { key: "timeout", label: "Timeout (seconds)", type: "text", readOnly: true },
-        { key: "configurationJson", label: "Action configuration", type: "textarea", readOnly: true },
-      ],
-      socketEvents: {
-        create: "triggerCreate",
-        update: "triggerUpdate",
-        delete: "triggerDelete",
-      },
-      defaultForm: {
-        name: "",
-        description: "",
-        projectId: null,
-        maxRetries: 3,
-        parallelLimit: 1,
-        timeout: 300,
-        triggerEventId: null,
-        triggerActionId: null,
-      },
-      triggerForm: {},
-      eventData: {},
-      actionData: {},
-      editingId: null,
-      viewFormData: null,
     };
   },
   computed: {
@@ -262,38 +96,11 @@ export default {
     },
     buttons() {
       return this.manageActions.map(
-        ({ handler, socketEvent, successToast, errorToast, modal, confirm, filter, ...btn }) => btn
+        ({ handler, socketEvent, confirm, ...btn }) => btn
       );
     },
     manageActionsByAction() {
       return Object.fromEntries(this.manageActions.map((a) => [a.action, a]));
-    },
-    selectedEvent() {
-      return this.$store.getters["table/trigger_event/getAll"].find(
-        (e) => e.id === this.triggerForm.triggerEventId && e.enabled && !e.deleted
-      );
-    },
-    selectedAction() {
-      return this.$store.getters["table/trigger_action/getAll"].find(
-        (a) => a.id === this.triggerForm.triggerActionId && a.enabled && !a.deleted
-      );
-    },
-    settingsFields() {
-      return this.resolveFormSchema(this.settingsFormSchema);
-    },
-    eventFields() {
-      return [this.resolveField(this.eventField)];
-    },
-    eventConfigFields() {
-      const schema = this.selectedEvent?.configuration?.formSchema || [];
-      return schema.map((field) => this.resolveField(field));
-    },
-    actionSelectFields() {
-      return [this.resolveField(this.actionField, { event: this.selectedEvent })];
-    },
-    actionConfigFields() {
-      const schema = this.selectedAction?.configuration?.formSchema || [];
-      return schema.map((field) => this.resolveField(field));
     },
     triggers() {
       const eventsById = Object.fromEntries(
@@ -323,76 +130,10 @@ export default {
           };
         });
     },
-    viewModalTitle() {
-      if (!this.viewFormData) return "Trigger";
-      return this.viewModalTitleTemplate.replace("{name}", this.viewFormData.name);
-    },
-  },
-  mounted() {
-    this.triggerForm = this.defaultTriggerForm();
   },
   methods: {
-    defaultTriggerForm() {
-      return {
-        ...this.defaultForm,
-        projectId: this.projectId,
-      };
-    },
-    resolveField(field, context = {}) {
-      if (field.options) return field;
-      const src = field.optionsSource;
-      if (!src) return field;
-
-      let rows = this.$store.getters[`table/${src.table}/getAll`].filter((r) => !r.deleted);
-
-      if (src.filter) {
-        rows = rows.filter((r) =>
-          Object.entries(src.filter).every(([k, v]) =>
-            Array.isArray(v) ? v.includes(r[k]) : r[k] === v
-          )
-        );
-      }
-
-      if (src.filterFromForm) {
-        rows = rows.filter((r) =>
-          Object.entries(src.filterFromForm).every(([k, formKey]) => {
-            const val = this.triggerForm[formKey];
-            return val == null || r[k] === val;
-          })
-        );
-      }
-
-      if (src.distinct) {
-        const seen = new Set();
-        rows = rows.filter((r) => {
-          const v = r[src.valueKey];
-          if (v == null || seen.has(v)) return false;
-          seen.add(v);
-          return true;
-        });
-      }
-
-      if (src.compatibleWithEvent && context.event) {
-        const provided = new Set(context.event.configuration?.provides || []);
-        rows = rows.filter((a) =>
-          (a.configuration?.requires || []).every((key) => provided.has(key))
-        );
-      }
-
-      const options = rows.map((r) => ({
-        name: src.labelKey === "configuration.label"
-          ? (r.configuration?.label || r.name)
-          : r[src.labelKey] || r[src.nameKey] || r.name,
-        value: r[src.valueKey],
-      }));
-
-      return {
-        ...field,
-        options: src.emptyOption ? [src.emptyOption, ...options] : options,
-      };
-    },
-    resolveFormSchema(schema) {
-      return schema.map((field) => this.resolveField(field));
+    openCreate() {
+      this.$refs.triggerStepper.openCreate(this.projectId);
     },
     onAction(data) {
       const row = data.params;
@@ -405,21 +146,12 @@ export default {
       const actionDef = this.manageActionsByAction[data.action];
 
       if (actionDef.handler === "viewModal") {
-        this.viewFormData = {
-          name: row.name,
-          eventLabel: row.eventLabel,
-          actionLabel: row.actionLabel,
-          maxRetries: String(row.maxRetries),
-          parallelLimit: String(row.parallelLimit),
-          timeout: String(row.timeout),
-          configurationJson: JSON.stringify(row.configuration || {}, null, 2),
-        };
-        this.$refs.viewModal.open();
+        this.$refs.triggerStepper.openView(row);
         return;
       }
 
       if (actionDef.handler === "editStepper") {
-        this.openEdit(row);
+        this.$refs.triggerStepper.openEdit(row);
         return;
       }
 
@@ -436,68 +168,6 @@ export default {
           }
         );
       }
-    },
-    openCreate() {
-      this.editingId = null;
-      this.triggerForm = this.defaultTriggerForm();
-      this.eventData = {};
-      this.actionData = {};
-      this.$refs.triggerStepper.open();
-    },
-    openEdit(row) {
-      this.editingId = row.id;
-      const config = row.configuration || {};
-      this.triggerForm = {
-        name: row.name,
-        description: config.description || "",
-        projectId: row.projectId || null,
-        maxRetries: row.maxRetries,
-        parallelLimit: row.parallelLimit,
-        timeout: row.timeout,
-        triggerEventId: row.triggerEventId,
-        triggerActionId: row.triggerActionId,
-      };
-      this.eventData = config.event || {};
-      this.actionData = config.action || {};
-      this.$refs.triggerStepper.open();
-    },
-    save() {
-      const payload = {
-        name: this.triggerForm.name,
-        triggerEventId: this.triggerForm.triggerEventId,
-        triggerActionId: this.triggerForm.triggerActionId,
-        projectId: this.triggerForm.projectId,
-        maxRetries: Number(this.triggerForm.maxRetries),
-        parallelLimit: Number(this.triggerForm.parallelLimit),
-        timeout: Number(this.triggerForm.timeout),
-        configuration: {
-          description: this.triggerForm.description,
-          event: this.eventData,
-          action: this.actionData,
-        },
-      };
-      const editing = this.editingId !== null;
-      const socketEvent = editing ? this.socketEvents.update : this.socketEvents.create;
-      if (editing) {
-        payload.id = this.editingId;
-      }
-      this.$socket.emit(socketEvent, payload, (res) => {
-        if (res.success) {
-          this.$refs.triggerStepper.close();
-          this.editingId = null;
-          this.eventBus.emit("toast", {
-            title: editing ? "Trigger updated" : "Trigger created",
-            message: `The trigger has been ${editing ? "updated" : "created"} successfully.`,
-            variant: "success",
-          });
-        } else {
-          this.eventBus.emit("toast", {
-            title: editing ? "Failed to update trigger" : "Failed to create trigger",
-            message: res.message || "Unknown error",
-            variant: "danger",
-          });
-        }
-      });
     },
   },
 };
